First Commit
This program works by entering one of the eight txt files in the program file directory into the arguments of the command line. The txt files are the data of the Hidato Puzzles. The program will then post a graphical image to user's screen and the puzzle will be solved.